Move inject_pyodide_bootstrap_cell to its own module
